rh - moved grade history fixtures to test directory
